fix(web): move normalizeName to its own module for react-refresh
