+++ apps/web/src/components/NewAgent/Steps/NameStep/normalize-name.ts
@@ -1,0 +1,9 @@
+export function normalizeName(input: string): string {
+  return input
+    .toLowerCase()
+    .trim()
+    .replace(/\s+/g, "-")
+    .replace(/[^a-z0-9-]/g, "")
+    .replace(/-+/g, "-")
+    .replace(/^-|-$/g, "");
+}
